Improve modularity of functions by abstracting away into toolkits
