+++ test/functions/test_google_maps_function.py
@@ -14,11 +14,7 @@
 from functools import wraps
 from unittest.mock import MagicMock, patch
 
-from camel.toolkits.google_maps_function import (
-    get_address_description,
-    get_elevation,
-    get_timezone,
-)
+from camel.toolkits.google_maps_toolkit import GoogleMapsToolkit
 
 
 def mock_googlemaps(test_func):
@@ -62,7 +58,7 @@ def test_get_address_description(mock_get, mock_client):
     mock_client.return_value = mock_instance
 
     # Call the function with a test address
-    result = get_address_description(
+    result = GoogleMapsToolkit().get_address_description(
         '1600 Amphitheatre Pk', region_code='US', locality='Mountain View'
     )
 
@@ -93,7 +89,7 @@ def test_get_elevation(mock_get, mock_client):
     mock_client.return_value = mock_instance
 
     # Call the function with a test latitude and longitude
-    result = get_elevation((40.71473, -73.99867))
+    result = GoogleMapsToolkit().get_elevation((40.71473, -73.99867))
 
     # Verify the result
     expected_result = (
@@ -121,7 +117,7 @@ def test_get_timezone(mock_get, mock_client):
     mock_client.return_value = mock_instance
 
     # Call the function with a test latitude and longitude
-    result = get_timezone(
+    result = GoogleMapsToolkit().get_timezone(
         (39.603481, -119.682251)
     )  # Coordinates for Los Angeles
 
@@ -140,10 +136,16 @@ def test_wrong_api_key(monkeypatch):
     monkeypatch.setenv('GOOGLEMAPS_API_KEY', 'invalid_api_key')
     expected_output = "Error: Invalid API key provided."
     assert (
-        get_address_description(
+        GoogleMapsToolkit().get_address_description(
             '1600 Amphitheatre Pk', region_code='US', locality='Mountain View'
         )
         == expected_output
     )
-    assert get_elevation((40.714728, -73.998672)) == expected_output
-    assert get_timezone((40.714728, -73.998672)) == expected_output
+    assert (
+        GoogleMapsToolkit().get_elevation((40.714728, -73.998672))
+        == expected_output
+    )
+    assert (
+        GoogleMapsToolkit().get_timezone((40.714728, -73.998672))
+        == expected_output
+    )
