+++ test/functions/test_twitter_function.py
@@ -13,11 +13,7 @@
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 from unittest.mock import MagicMock, patch
 
-from camel.toolkits.twitter_function import (
-    create_tweet,
-    delete_tweet,
-    get_my_user_profile,
-)
+from camel.toolkits import TwitterToolkit
 
 
 def test_create_tweet(monkeypatch):
@@ -40,14 +36,14 @@ def test_create_tweet(monkeypatch):
     monkeypatch.setattr('builtins.print', lambda x: captured_output.append(x))
 
     # Use patch to mock the get_oauth_session method
-    patch_path = 'camel.toolkits.twitter_function.get_oauth_session'
+    patch_path = 'camel.toolkits.twitter_toolkit.get_oauth_session'
     with patch(patch_path) as mock_get_oauth_session:
         # Configure the mock OAuth session's post method
         # to return the mock response object
         mock_get_oauth_session.return_value.post.return_value = mock_response
 
         # Call the create_tweet function
-        response = create_tweet(text="Test tweet.")
+        response = TwitterToolkit().create_tweet(text="Test tweet.")
 
         # Verify the output
         expected_start = (
@@ -76,14 +72,14 @@ def test_delete_tweet(monkeypatch):
     monkeypatch.setattr('builtins.print', lambda x: captured_output.append(x))
 
     # Use patch to mock the get_oauth_session method
-    patch_path = 'camel.toolkits.twitter_function.get_oauth_session'
+    patch_path = 'camel.toolkits.twitter_toolkit.get_oauth_session'
     with patch(patch_path) as mock_get_oauth_session:
         # Configure the mock OAuth session's delete method
         # to return the mock response object
         mock_get_oauth_session.return_value.delete.return_value = mock_response
 
         # Call the delete_tweet function
-        response = delete_tweet(tweet_id="11111")
+        response = TwitterToolkit().delete_tweet(tweet_id="11111")
 
         # Verify the output
         expected_start = (
@@ -131,7 +127,7 @@ def test_get_user_me(monkeypatch):
     }
 
     # Use patch to mock the get_oauth_session method
-    patch_path = 'camel.toolkits.twitter_function.get_oauth_session'
+    patch_path = 'camel.toolkits.twitter_toolkit.get_oauth_session'
     with patch(patch_path) as mock_get_oauth_session:
         mock_response = MagicMock()
         mock_response.json.return_value = mock_json_response
@@ -139,7 +135,7 @@ def test_get_user_me(monkeypatch):
         mock_get_oauth_session.return_value.get.return_value = mock_response
 
         # Call the get_user_me function
-        response = get_my_user_profile()
+        response = TwitterToolkit().get_my_user_profile()
 
         # Verify the returned user information report
         expected_report = (
